chore: update, fix test
